requirements.txt
numpy==1.19.5
picamera2==0.3.3
Pillow==9.3.0
tflite_runtime==2.10.0
